Add a fallback cli tool in Python to derive Bitcoin addresses on 32 bits systems (#37)
* added a python script for xpub derivation on 32-bits

* install pip too

* first working draft for python fallback on php 32 bits systems

* first working draft for python fallback on php 32 bits systems

* this fixes a bug on 32 bits systems

* this fixes a bug on 32 bits systems

* added tests for new fallback component

* mark test as skipped if its not present or non existent for travis

* code smells

* updated test bootstrap

+++ src/Client/BitvavoClient.php
@@ -38,7 +38,7 @@ class BitvavoClient implements BitvavoClientInterface
         int $now = null
     ): array {
         if (null === $now) {
-            $now = time() * 1000;
+            $now = (int) (time() * 1000);
         }
 
         $query = http_build_query($parameters, '', '&');
